requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
paramiko
python-dotenv
openai
pywinrm
pydantic
pyvmomi
requests